requirements.txt
nmigen @ git+https://github.com/nmigen/nmigen.git
nmigen-boards @ git+https://github.com/alanvgreen/nmigen-boards.git
attrs
